Food Delivery related changes added in tenant locations table + Modules table dropped in favour of TenantModuleEnum + minor fixes Imported from Changeset 2965

+++ GaneEcommerce.Services/Tenants/TenantsServices.cs
@@ -76,7 +76,7 @@ namespace Ganedata.Core.Services
 
         public bool IsModuleEnabled(int tenantId, TenantModuleEnum tenantModule)
         {
-            return _currentDbContext.TenantModules.Any(m => m.ModuleId == (int)tenantModule && m.TenantId == tenantId);
+            return _currentDbContext.TenantModules.Any(m => m.ModuleId == tenantModule && m.TenantId == tenantId);
         }
 
     }
